+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Only partial changelog, [commit history](https://framagit.org/Shnoulle/LimeSurve
 
 ## unreleased
 
+### Feature
+- Allow to update via cli
+- Allow to update runtime directory in config (LimeSurvey API 3 compatible only)
+
 ### [1.2.2] - 2016-01-12
 - Update [PHPMailer](https://github.com/PHPMailer/PHPMailer/releases/tag/v5.2.22) for security issue